Random State
In sklearn.model_selection.train_test_split,
parameter 'random_state' can be passed.
It is important to set a random state
to make sure the result is reproducible.

The seed of randomness '42' does not contain any specific meaning,
but derived from the 'Hitchhiker's Guide to the Galaxy',
by Douglas Adams.
According to the literature, it is considered as
'the Answer to the Ultimate Question of Life, the Universe
and Everything' by the Deep Thought.